zalo/Zalo/requirements.txt
airtest==1.1.1
androidviewclient==15.8.1
Appium-Python-Client==0.48
asgiref==3.2.3
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
certifi==2019.9.11
cffi==1.13.2
channels==2.3.1
chardet==3.0.4
comtypes==1.1.7
concurrent-log-handler==0.9.16
ConcurrentLogHandler==0.9.1
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
daphne==2.3.0
DBUtils==1.3
Django==2.2.6
djangorestframework==3.10.3
dwebsocket==0.5.11
facebook-wda==0.2.1
func-timeout==4.3.5
hrpc==1.0.8
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
MouseInfo==0.1.2
mss==4.0.3
mysqlclient==1.4.4
numpy==1.17.3
opencv-contrib-python==3.4.2.17
pika==1.1.0
Pillow==6.2.1
pocoui==1.0.78
portalocker==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAutoGUI==0.9.48
pycparser==2.19
PyGetWindow==0.0.7
PyHamcrest==1.9.0
PyMsgBox==1.0.7
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyperclip==1.7.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.25
python-dateutil==2.8.1
python-poco==0.1.0
PyTweening==1.0.3
pytz==2019.3
pywin32==225
pywinauto==0.6.3
redis==3.3.11
requests==2.22.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.6
websocket-client==0.48.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
